v2.0.0-beta.10

+++ build.js
@@ -12,7 +12,7 @@ var spinner = ora('building ' +
   (process.env.BUILD == 'publish' ? 'vonic for production' : 'gh-pages') + '...')
 spinner.start()
 
-shell.rm('-rf', './dist')
+if (process.env.BUILD == 'publish') shell.rm('-rf', './dist')
 
 webpack(webpackConfig, function (err, stats) {
   spinner.stop()
